engine: first semi-working console commit
still a lot to do, but this is a proof of concept

+++ src/engine/coord/decl.h
@@ -49,7 +49,7 @@ struct camhud;
 struct camhud_delta;
 
 /* terminal character position-based types */
-using term_t = uint16_t;
+using term_t = int;
 
 struct term;
 struct term_delta;
